requirements.txt
matplotlib
librosa
pyaudio
pydub
opencv-python
PyOpenGL
PyOpenGL-accelerate
pyglut
pyglet
pygame
sounddevice
pyrr